Add a generic load_check method
This commit adds `Config::Identity->load_check` for generic use
to load a stub file and check for a required set of keys.

    Config::Identity->load_check( $stub, [qw/user pass/]);

    Config::Identity->load_check( $stub, sub { ... } );

This will allow Config::Identity to be used for simple cases other
than PAUSE and GitHub without requiring a subclass or requiring users
to write their own generic missing key checks.

This commit does not rewrite the PAUSE or GitHub subclasses in terms
of this method, as those classes provide independent methods for loading
and checking or both and it's better that those be self-consistent.